chore: Clean up several imports and definitions
Improves import hygiene, and transfers some duplicate code into different
util modules.
